Refactor the project - Extracted Search, Sort, and related logic into separate components. - Moved those component into separate files: `Search.jsx`, `Sort.jsx`, `PluginList.jsx`. - Added project states in Contexts folder. - Extracted the styles of `App.css` into the corresponding components.
