+++ aiida-registry-app/src/Components/MainIndex.css
@@ -1,0 +1,83 @@
+.globalsummary-box {
+    width:90%;
+    padding:10px;
+    margin: 20px;
+    padding-left: 40px;
+    padding-right: 40px;
+    background-color: rgba(5,115,255,0.19);
+    border:1px solid rgba(0,65,127,0.4);
+    border-radius:4px;
+    -webkit-box-shadow:inset 0 1px 1px rgba(0,0,0,.05);
+    box-shadow:inset 0 1px 1px rgba(0,0,0,.05);
+    line-height: 1.5;
+}
+
+main {
+    padding-left: 16px;
+    margin-left: 16px;
+    padding-top: 0px;
+    margin-top: 0px;
+    margin-right: 16px;
+    padding-right: 0px;
+    padding-bottom: 0px;
+    margin-bottom: 0px;
+    max-width: 1280px;
+    margin: 0 auto;
+  }
+
+.bar-container {
+    display: flex;
+    flex-direction: row;
+    margin: 0 2%;
+}
+
+.summaryinfo {
+  color: #000;
+  font-size: 80%;
+  margin-bottom: 12px;
+  margin-top: 12px;
+}
+
+ul.plugin-info {
+list-style: none;
+margin-left: 0;
+padding-left: 0;
+  padding-top: 5px;
+}
+
+ul.plugin-info li {
+padding-left: 1em;
+text-indent: -1em;
+}
+
+ul.plugin-info li:before {
+content: "→";
+padding-right: 5px;
+}
+
+@media only screen and (min-width : 150px) and (max-width : 780px)
+  {
+    .globalsummary-box
+      {
+        width:90%;
+        padding:1px;
+        margin: 20px 0px 20px 0px;
+        padding-left: 4px;
+        padding-right: 4px;
+      }
+
+    .globalsummary-box .badge
+      {
+        float: left;
+      }
+
+    .bar-container
+    {
+    display: block;
+    }
+
+    #entrylist h1 {
+        font-size: 25px;
+      }
+
+}
